+++ pusher/http.py
@@ -43,6 +43,8 @@ def make_query_string(params):
 def process_response(status, body):
     if status == 200:
         return json.loads(body)
+    if status == 202:
+        return True
     elif status == 400:
         raise PusherBadRequest(body)
     elif status == 401:
